interacts with page data
