+++ packages/edp-form-extension/package.json
@@ -16,7 +16,7 @@
     "@citric/icons": "5.9.0",
     "@citric/ui": "6.5.5",
     "@stack-spot/portal-components": "2.24.1",
-    "@stack-spot/portal-extension": "0.14.3",
+    "@stack-spot/portal-extension": "0.16.0",
     "@stack-spot/portal-theme": "1.1.0",
     "@stack-spot/portal-translate": "1.2.2",
     "@tanstack/react-query": "~5.59.15",
@@ -35,6 +35,7 @@
     "@typescript-eslint/eslint-plugin": "^6.10.0",
     "@typescript-eslint/parser": "^6.10.0",
     "@vitejs/plugin-react": "^4.2.0",
+    "@stack-spot/portal-network": "0.158.0",
     "bestzip": "^2.2.1",
     "eslint": "^8.53.0",
     "eslint-plugin-filenames": "^1.3.2",
